code: Dump the video framebuffer to the console.
This is ASCII-DOOM into the console! The framebuffer is converted
from a 256 colors palette to a 16 ascii-characters palette.

NOTE: It seems printing colors to the console is time consuming,
resulting in a non-fluid refresh rate. Instead of printing 16
colors (ANSI) symbols to the console, the palette is switched to
black & white symbols, trying to create the illusion of a grayscale.