code: Remove the X11 legacy renderer.
Remove all calls/references to X11 and shm, let a clean skeleton
of functions that must be implemented to get a the rendering of the
video framebuffer.

The video framebuffer is still allocated in src/i_video.c, and with
malloc(). Might worth considering changing this.

Control support (initially handled with X11) has been also removes,
look at the original sources to understand how to reimplement this.
The screen-multiplying has also been removed, what's the intended
use of this feature?